+++ src/webserver/database/mod.rs
@@ -2,7 +2,7 @@ mod connect;
 mod csv_import;
 pub mod execute_queries;
 pub mod migrations;
-pub mod mime_detection;
+pub mod blob_to_data_url;
 mod sql;
 mod sqlpage_functions;
 mod syntax_tree;
