+++ src/webserver/database/sqlpage_functions/functions.rs
@@ -1,7 +1,7 @@
 use super::RequestInfo;
 use crate::webserver::{
     database::{
-        execute_queries::DbConn, sql_to_json::vec_to_data_uri_with_mime,
+        blob_to_data_url::vec_to_data_uri_with_mime, execute_queries::DbConn,
         sqlpage_functions::url_parameter_deserializer::URLParameters,
     },
     http::SingleOrVec,
